10_lesson/requirements.txt
selenium>=4.0.0
pytest>=7.0.0
pytest-xdist
allure-pytest>=2.9.0
webdriver-manager>=3.8.0